Added fix for V6 contents format. showDocument now works.
Added fix for V6 contents format. Pages that are deleted still appear in cPages.pages but with a deleted key. I now filter out those entries.
Added getAvailablePages which will determine the file ids that will be sent to the client when requesting a document view.
 - It reads from the rm_fs pages entry for the uuid to get the current page ids, checks that the thumbnails of each page are there as well as the svg images. The returned value includes the file for the thumb and the svg.
 - If the svg is not included for a page then the thumbnail will be used instead by the client.
When the client does an rsync it will also be sent the new rm_fs file to updated itself with. This also reloads the file nav.
The showDocument function is now working. One page view with the thumbnails available to click on to jump directly to that page.