Thumbs show the active page, page displayed will be the newer file (thumb or svg).
Thumbs can be hovered and are shown highlighted after selecting.
updatePage will now display whichever file (thumb or svg) is the newer file.
getAvailablePages will determine which is newer (thumb or svg).
createThumb will not try to add the thumb image if the thumb is found to be part of the "missing" array. This results in blank thumbnails but no error.
getItemsInDir now returns filepath, mtimeMs, and ext.