Documents will always have thumbnails because they are created from the .svg pages now.
Added background color to the document view thumbs and the file nav thumbs.
Document view thumbs will display the .jpg thumb from the device or the .png thumb generated when the document is processed. Whichever is newer.
Added a ssh connection check to the BASH scripts that use ssh or rsync.
Wrote createThumbs.sh which will use the .svg files to create smaller dimensioned .png files using rsvg-convert which are then optimized by pngquant to a smaller filesize.
run_fullDownloadAndProcessing now creates the svgThumbs which much smaller .png file versions of the .svg pages.
getAvailablePages will now also determine with thumb is newer, the .jpg or the .png.