Refactors, error handling, local document deletion.
Improve error handling in convert.
Improve error handling in runCommand_exec_progress.
In m_sync, files in the trash or marked as deleted will not be added to needsUpdate.json.
rsyncUpdate will now delete local document folders if the document is no longer on the device.
With PDF downloads I added a check to make sure that the USB web interface is active.
Added support for deleting files on the device that are marked as deleted. This is WIP and not wired into any interface yet.
Web interface can now show the contents of "trash" and "deleted".
Refactor getParentPathBreadcrumbs.
Refactor getEntriesInCollectionType.