Added diskFree, recent documents, sidebar updates.
Added new script that will return the storage space available on the device (sync_getDiskUsage.sh).
Updates to the sidebar for fileNav.
 - You can now select from the last 20 opened or modified documents as well as from favorites.
  - These values will update on page load and on sync.
 - Shows the storage space used on the device.
 - Can switch the view to "My files", "trash", and "deleted".
Fixed bug in getAvailablePages that can trigger if the thumbnails folder does not exist for a document. (After the 3.2 update it appears that the thumbnail folders disappeared and that is when the bug became apparent.)